# docs/openapi.yml
openapi: 3.0.3
info:
  title: SmartTeach API
  version: 0.2.0
  description: |-
    Актуальная спецификация REST API для MVP проекта SmartTeach.
    Спецификация обновлена после пересмотра предметной области и синхронизирована
    с новой структурой базы данных, зафиксированной в `docs/07-database-structure.md`.

    Основные принципы текущего API:
    - центральная сущность системы — группа;
    - `class` и `course` в MVP не выделяются;
    - у группы есть отдельные настройки функций;
    - вступление в группу зависит от `accessMode`;
    - используются локальная авторизация и refresh-сессии;
    - поддерживаются групповые и личные чаты `1 на 1`;
    - расписание строится из уроков, дедлайнов заданий и кастомных событий.
servers:
  - url: http://localhost:3000/api/v1
    description: Локальный REST API
...
paths:
...
  /groups:
    get:
      description: Возвращает публичные группы без входа в аккаунт и дополняет ответ данными участия, если передан access token.
      parameters:
        - in: query
          name: search
          required: false
          schema:
            example: patterns
            type: string
        - in: query
          name: code
          required: false
          schema:
            example: WEBSPR1
            type: string
        - in: query
          name: accessMode
          required: false
          schema:
            enum:
              - OPEN
              - BY_REQUEST
              - CLOSED
            type: string
        - in: query
          name: status
          required: false
          schema:
            enum:
              - ACTIVE
              - ARCHIVED
              - DELETED
            type: string
        - in: query
          name: joinedOnly
          required: false
          schema:
            example: true
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Group'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      summary: Получить список групп
      tags:
        - Groups
    post:
      description: Создает группу, настройки группы и запись владельца в составе участников.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGroupRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Group'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Создать группу
      tags:
        - Groups
...
  /groups/{groupId}/assignments:
    get:
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: query
          name: status
          required: false
          schema:
            enum:
              - DRAFT
              - PUBLISHED
              - ARCHIVED
            type: string
        - in: query
          name: lessonId
          required: false
          schema:
            format: uuid
            example: 77777777-7777-4777-8777-777777777777
            type: string
        - in: query
          name: materialSectionId
          required: false
          schema:
            format: uuid
            example: 88888888-8888-4888-8888-888888888888
            type: string
        - in: query
          name: materialSubsectionId
          required: false
          schema:
            format: uuid
            example: 99999999-9999-4999-8999-999999999999
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Assignment'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Получить список заданий группы
      tags:
        - Assignments
    post:
      description: Доступно владельцу и администраторам группы, если assignments_enabled включен.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAssignmentRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Создать задание
      tags:
        - Assignments
  /groups/{groupId}/assignments/{assignmentId}:
    get:
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: assignmentId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Получить задание
      tags:
        - Assignments
    patch:
      description: Позволяет менять статус, дедлайн, lessonId и набор привязанных файлов задания.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: assignmentId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAssignmentRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Assignment'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Обновить задание
      tags:
        - Assignments
  /groups/{groupId}/assignments/{assignmentId}/submissions:
    get:
      description: Обычный участник видит только собственные submissions. OWNER и ADMIN могут видеть все submissions задания.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: assignmentId
          required: true
          schema:
            format: uuid
            type: string
        - in: query
          name: mineOnly
          required: false
          schema:
            example: true
            type: boolean
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Submission'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Получить попытки сдачи задания
      tags:
        - Assignments
    post:
      description: Создает новую попытку текущего пользователя и автоматически назначает следующий attemptNumber.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: assignmentId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateSubmissionRequest'
      responses:
        '201':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Создать новую попытку сдачи
      tags:
        - Assignments
  /groups/{groupId}/assignments/{assignmentId}/submissions/{submissionId}:
    get:
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: assignmentId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: submissionId
          required: true
          schema:
            format: uuid
            type: string
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Получить попытку сдачи
      tags:
        - Assignments
    patch:
      description: Используется как для редактирования своей попытки, так и для проверки работы manager-ролями.
      parameters:
        - in: path
          name: groupId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: assignmentId
          required: true
          schema:
            format: uuid
            type: string
        - in: path
          name: submissionId
          required: true
          schema:
            format: uuid
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSubmissionRequest'
      responses:
        '200':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Submission'
        '400':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '403':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: ''
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
      security:
        - bearerAuth: []
      summary: Обновить попытку сдачи
      tags:
        - Assignments
...
components:
  securitySchemes:
    bearerAuth:
      scheme: bearer
      bearerFormat: JWT
      type: http
  schemas:
    Assignment:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 88888888-8888-4888-8888-888888888888
        groupId:
          type: string
          format: uuid
          example: 66666666-6666-4666-8666-666666666666
        lessonId:
          type: object
          format: uuid
          nullable: true
          example: 77777777-7777-4777-8777-777777777777
        materialSectionId:
          type: object
          format: uuid
          nullable: true
          example: 88888888-8888-4888-8888-888888888888
        materialSubsectionId:
          type: object
          format: uuid
          nullable: true
          example: 99999999-9999-4999-8999-999999999999
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: Prepare contract review
        content:
          type: object
          nullable: true
          example: Подготовьте разбор API-контракта и зафиксируйте спорные поля.
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
          example: PUBLISHED
        dueAt:
          type: object
          format: date-time
          nullable: true
        maxScore:
          type: object
          minimum: 0
          nullable: true
          example: 100
        publishedAt:
          type: object
          format: date-time
          nullable: true
        archivedAt:
          type: object
          format: date-time
          nullable: true
        createdByUserId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileObject'
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - groupId
        - title
        - status
        - createdByUserId
        - files
        - createdAt
        - updatedAt
...
    CreateAssignmentRequest:
      type: object
      properties:
        lessonId:
          type: string
          format: uuid
          example: 77777777-7777-4777-8777-777777777777
        materialSectionId:
          type: string
          format: uuid
          example: 88888888-8888-4888-8888-888888888888
        materialSubsectionId:
          type: string
          format: uuid
          example: 99999999-9999-4999-8999-999999999999
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: Prepare contract review
        content:
          type: string
          example: Сравните итоговую реализацию с OpenAPI и опишите расхождения.
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
          example: DRAFT
        dueAt:
          type: string
          format: date-time
          example: '2026-04-20T18:00:00.000Z'
        maxScore:
          type: number
          minimum: 0
          example: 100
        fileIds:
          example:
            - aaaaaaa4-aaaa-4aaa-8aaa-aaaaaaaaaaa4
          type: array
          items:
            type: string
      required:
        - title
...
    CreateGroupRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 150
          example: API Patterns Lab
        description:
          type: string
          example: Группа для разбора компонентного мышления и интерфейсных паттернов.
        accessMode:
          type: string
          enum:
            - OPEN
            - BY_REQUEST
            - CLOSED
          example: BY_REQUEST
        settings:
          $ref: '#/components/schemas/GroupSettings'
      required:
        - name
        - accessMode
        - settings
...
    CreateSubmissionRequest:
      type: object
      properties:
        text:
          type: string
          example: Готовая работа и комментарии по спорным местам реализации.
        fileIds:
          example:
            - bbbbbbb1-bbbb-4bbb-8bbb-bbbbbbbbbbb1
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - DRAFT
            - SUBMITTED
          example: DRAFT
    ErrorResponse:
      type: object
      properties:
        statusCode:
          type: number
          example: 400
        message:
          type: string
          example: Validation failed
        errors:
          example:
            - 'email: Invalid email address'
          type: array
          items:
            type: string
      required:
        - statusCode
        - message
...
    Group:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 66666666-6666-4666-8666-666666666666
        code:
          type: string
          example: WEBSPR1
        name:
          type: string
          minLength: 2
          maxLength: 150
          example: Web Basics Spring 2026
        description:
          type: object
          nullable: true
          example: Открытая группа для отработки HTML, CSS и базового JavaScript.
        ownerId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        owner:
          $ref: '#/components/schemas/PublicUser'
        accessMode:
          type: string
          enum:
            - OPEN
            - BY_REQUEST
            - CLOSED
          example: OPEN
        status:
          type: string
          enum:
            - ACTIVE
            - ARCHIVED
            - DELETED
          example: ACTIVE
        settings:
          $ref: '#/components/schemas/GroupSettings'
        membersCount:
          type: number
          minimum: 0
          example: 3
        viewerMembershipRole:
          type: string
          enum:
            - OWNER
            - ADMIN
            - USER
          nullable: true
          description: Роль текущего пользователя в группе, если он уже состоит в ней.
          example: USER
        viewerJoinRequestStatus:
          type: string
          enum:
            - PENDING
            - APPROVED
            - REJECTED
          nullable: true
          description: Последний известный статус заявки текущего пользователя в эту группу.
          example: PENDING
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        archivedAt:
          type: object
          format: date-time
          nullable: true
        deletedAt:
          type: object
          format: date-time
          nullable: true
      required:
        - id
        - code
        - name
        - ownerId
        - owner
        - accessMode
        - status
        - settings
        - membersCount
        - createdAt
        - updatedAt
...
    PublicUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        displayName:
          type: string
          example: Alex Teacher
        bio:
          type: object
          nullable: true
          example: Преподает веб-разработку и ведет открытые учебные группы.
        avatarFileId:
          type: object
          format: uuid
          nullable: true
          example: aaaaaaa1-aaaa-4aaa-8aaa-aaaaaaaaaaa1
        avatarUrl:
          type: object
          format: uri
          nullable: true
          example: http://localhost:9000/smarteach-files/avatars/2026/04/12/aBcD1234xYz9.png?X-Amz-Algorithm=AWS4-HMAC-SHA256
      required:
        - id
        - displayName
...
    Submission:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 99999999-9999-4999-8999-999999999999
        assignmentId:
          type: string
          format: uuid
          example: 88888888-8888-4888-8888-888888888888
        authorId:
          type: string
          format: uuid
          example: 11111111-1111-4111-8111-111111111111
        attemptNumber:
          type: number
          minimum: 1
          example: 2
        text:
          type: object
          nullable: true
          example: Финальная версия работы.
        status:
          type: string
          enum:
            - DRAFT
            - SUBMITTED
            - REVIEWED
          example: SUBMITTED
        files:
          type: array
          items:
            $ref: '#/components/schemas/FileObject'
        score:
          type: object
          minimum: 0
          nullable: true
          example: 90
        feedback:
          type: object
          nullable: true
          example: Хорошая работа, но нужно закрыть edge case с правами доступа.
        submittedAt:
          type: object
          format: date-time
          nullable: true
        reviewedByUserId:
          type: object
          format: uuid
          nullable: true
          example: 22222222-2222-4222-8222-222222222222
        reviewedAt:
          type: object
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        author:
          $ref: '#/components/schemas/PublicUser'
        reviewer:
          nullable: true
          allOf:
            - $ref: '#/components/schemas/PublicUser'
      required:
        - id
        - assignmentId
        - authorId
        - attemptNumber
        - status
        - files
        - createdAt
        - updatedAt
        - author
...
    UpdateAssignmentRequest:
      type: object
      properties:
        lessonId:
          type: object
          format: uuid
          nullable: true
          example: 77777777-7777-4777-8777-777777777777
        materialSectionId:
          type: object
          format: uuid
          nullable: true
          example: 88888888-8888-4888-8888-888888888888
        materialSubsectionId:
          type: object
          format: uuid
          nullable: true
          example: 99999999-9999-4999-8999-999999999999
        title:
          type: string
          minLength: 2
          maxLength: 200
          example: Updated assignment title
        content:
          type: string
          example: Уточните формат финального отчета и дедлайн.
        status:
          type: string
          enum:
            - DRAFT
            - PUBLISHED
            - ARCHIVED
          example: ARCHIVED
        dueAt:
          type: string
          format: date-time
          example: '2026-04-22T18:00:00.000Z'
        maxScore:
          type: number
          minimum: 0
          example: 80
        fileIds:
          example:
            - aaaaaaa4-aaaa-4aaa-8aaa-aaaaaaaaaaa4
          type: array
          items:
            type: string
...
    UpdateSubmissionRequest:
      type: object
      properties:
        text:
          type: string
          example: Обновленная работа с исправлением замечаний.
        fileIds:
          example:
            - bbbbbbb1-bbbb-4bbb-8bbb-bbbbbbbbbbb1
          type: array
          items:
            type: string
        status:
          type: string
          enum:
            - DRAFT
            - SUBMITTED
            - REVIEWED
          example: REVIEWED
        score:
          type: number
          minimum: 0
          example: 95
        feedback:
          type: string
          example: Хорошая структура ответа, но добавьте тест на отказ в доступе.
...